requirements.txt
bert_score==0.3.13
click==8.2.1
hydra-core==1.3.2
numpy==1.26.0
omegaconf==2.3.0
openai==1.107.3
pandas==2.3.2
sacrebleu==2.5.1
setuptools==75.8.0
tiktoken==0.11.0
tqdm==4.67.1
transformers==4.45.2
vllm==0.6.6
ray==2.49.1